+++ modules/libcom/src/osi/os/RTEMS-score/osdReadline.c
@@ -16,6 +16,8 @@
 #include <string.h>
 #include <libtecla.h>
 
+#include "epicsReadlinePvt.h"
+
 struct osdContext {};
 
 /*
